server search complete

+++ src/com/fullsink/mp/WebServer.java
@@ -5,6 +5,7 @@ import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
+import java.io.InputStream;
 import java.net.InetSocketAddress;
 import java.net.InetAddress;
 import java.net.UnknownHostException;
@@ -13,9 +14,10 @@ import java.util.Collection;
 import org.java_websocket.WebSocket;
 import org.java_websocket.handshake.ClientHandshake;
 import org.java_websocket.server.WebSocketServer;
+
+
 import static com.fullsink.mp.Const.*;
 
-import android.content.res.AssetFileDescriptor;
 import android.util.Base64;
 
 
@@ -118,8 +120,10 @@ public class WebServer extends WebSocketServer {
         	
         	String[] xdump = mnact.getFilesDir().list();
     		for (int i=0; i< xdump.length; i++){
-    			System.out.println("Dump of cued bits : "+xdump[i]+"  : "+i);
-    			new File(mnact.getFilesDir(),xdump[i]).delete();
+//    			System.out.println("Dump of cued bits : "+xdump[i]+"  : "+i);
+    			if (!xdump[i].equals(HTMLDIR)) {
+    				new File(mnact.getFilesDir(),xdump[i]).delete();
+    			}
        		}
         }
         
@@ -131,7 +135,7 @@ public class WebServer extends WebSocketServer {
  
         	try {
         		
-        		byte [] xbuf = new byte[65536];     		
+        		byte [] xbuf = new byte[BASE_BLOCKSIZE];     		
         		      		
            		//Parent directories need to be generated first
         		
@@ -201,7 +205,6 @@ public class WebServer extends WebSocketServer {
     	   }
        }
        
-
        
        private class FileServer implements Runnable {
        	
@@ -274,6 +277,54 @@ public class WebServer extends WebSocketServer {
     	       }
        }
  
+       
+        public void initHTML() {
+        	
+       	// Put this back later after testing complete
+//       	if (! new File(mnact.getFilesDir(),HTMLDIR).isDirectory()) {
+
+       		mnact.textOut("In initHTML");
+        	 
+        	try {
+        		
+        		byte [] xbuf = new byte[BASE_BLOCKSIZE];  
+        		File htmldest,htmlpar;
+        		      		
+           		//Parent directories need to be generated first
+        		String[] afiles = mnact.getAssets().list("html");
+
+        			htmlpar = new File(mnact.getFilesDir(),HTMLDIR);
+        			htmlpar.mkdirs();
+
+        		// Copy contents of assets over to files
+        		for (int i=0; i<afiles.length; i++){
+        			System.out.println("Transfer : "+afiles[i]); 
+        			
+            		htmldest = new File(htmlpar, afiles[i]);
+            		htmldest.createNewFile();
+        		
+            		InputStream in = mnact.getAssets().open("html/"+afiles[i]);
+ //       		System.out.println("Create input");        	    
+        	    OutputStream out = new FileOutputStream(htmldest);
+        		System.out.println("Create output");
+        	    
+        	    
+        	    // Transfer bytes from in to out
+        	    int len;
+        	    while ((len = in.read(xbuf)) > 0) {
+        	        out.write(xbuf, 0, len);
+            		System.out.println("Transfer output");
+        	    }
+        	    in.close();
+        	    out.close();
+        		}
+        	    
+        	} catch (IOException e) {
+        		System.out.println( "File I/O error " + e);
+        	}
+
+ //      	}
+        }
 }
 
 /*
